Changed logging name to not conflict with standard library Added documentation around supporting Added instructions for hatch setup and integration in vscode Other minor fixes related to linting
