+++ src/kagglehub/__init__.py
@@ -1,6 +1,6 @@
 __version__ = "0.2.4"
 
-import kagglehub.logging  # configures the library logger.
+import kagglehub.logger  # configures the library logger.
 from kagglehub import colab_cache_resolver, http_resolver, kaggle_cache_resolver, registry
 from kagglehub.auth import login
 from kagglehub.models import model_download, model_upload
